backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
httpx==0.27.0
python-multipart==0.0.9

# Article scraping
newspaper3k==0.2.8
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0

# NLP / AI
transformers==4.40.0
torch==2.2.2
sentencepiece==0.2.0

# Utilities
python-dateutil==2.9.0
